ATH-01-001: resolve path and check start before loading blueprints
This is even less of an issue since 411ef239f63e4d3beacd8297d4be54b29fb30127, since with that commit we only allow files that the listing returns
