+++ authentik/blueprints/models.py
@@ -82,7 +82,9 @@ class BlueprintInstance(SerializerModel, ManagedModel, CreatedUpdatedModel):
     def retrieve_file(self) -> str:
         """Get blueprint from path"""
         try:
-            full_path = Path(CONFIG.y("blueprints_dir")).joinpath(Path(self.path))
+            full_path = Path(CONFIG.y("blueprints_dir")).joinpath(Path(self.path)).resolve()
+            if not str(full_path).startswith(CONFIG.y("blueprints_dir")):
+                raise BlueprintRetrievalFailed("Invalid blueprint path")
             with full_path.open("r", encoding="utf-8") as _file:
                 return _file.read()
         except (IOError, OSError) as exc:
